cdl: improve human readable t2 printing
After uploading a page with t2cdlunits == 0, cdladm prints something
like this:

  Descriptor 7:
    max inactive time        : no limit
    max active time          : 0 ns
    max active policy        : abort
    duration guideline       : no limit

This doesn't look very nice.

According to spc6:
"If the T2CDLUNITS field is set to 0h, the MAX INACTIVE TIME field shall
be ignored."
"If the T2CDLUNITS field is set to 0h, the MAX ACTIVE TIME field shall
be ignored."
"If the T2CDLUNITS field is set to 0h, the COMMAND DURATION GUIDELINE
field shall be ignored."

Since the limit is ignored by the device when t2cdlunits == 0,
simply print "no limit" in this case (and avoid printing the
policy, since without a limit, the policy will never by applied).

Signed-off-by: Niklas Cassel <niklas.cassel@wdc.com>