cdl: fix off-by-one in print outs
If you have an error in the data beneath e.g.
== descriptor: 7
then you expect that the error message will refer to the
actual descriptor where the error is.

Most prints already add one to the index before printing.
Follow the existing convention and fix the few prints that
do not add one before printing.

Signed-off-by: Niklas Cassel <niklas.cassel@wdc.com>