cdl: scsi: silence MAINTENANCE_IN errors
For disks and HBAs not supporting CDLs, the MAINTENANCE IN command to
issue a REPORT_SUPPORTED_OPERATION_CODES service action is not
necessarilly supported. executing cdl_scsi_init() then leads to many
error messages that are not relevant. Silence these messages unless the
verbose option is specified.

Note that the detailed error messages for any failed command in
cdl_cmd_exec() are also silenced unless the verbose option is specified.

Signed-off-by: Damien Le Moal <damien.lemoal@opensource.wdc.com>