tests: verify that fio does not run longer than expected
If there is an I/O timeout, then the Linux driver will hard reset the HBA
after 30 seconds.

We have tests that are expected to fail. These tests are supposed to run
less than the runtime supplied on the fio command line.

We have tests that are not expected to fail. These tests are supposed to
run the whole runtime supplied on the fio command line.

Regardless, verify that fio does not run longer than expected.

A small 2 second buffer is added to account for the time it takes to start
and exit the fio binary.

Add a check so that we cannot accidentally PASS a test case that runs
longer than expected.

Signed-off-by: Niklas Cassel <niklas.cassel@wdc.com>