tests: improve test run time logic
If we expect error, we want to have a short run time. We should still get an
error in this shorter run time. Having a longer test time when we expect an
error is actually LESS reliable, as we might get the error first after e.g.
59 seconds (when the run time would be 60 seconds).

When we don't expect error, we want to have a long run time, such that we can
be more confident that we do not encounter an error when it is not expected.

Signed-off-by: Niklas Cassel <niklas.cassel@wdc.com>