tests: Fix inactive time non-ncq read tests
When executing read commands without NCQ enabled, a read command is
issued to the drive after the completion of the previous command (QD=1
operation). There is thus never any inactive time for the command as it
is executed immediately by the device. We should thus not expect any
ETIME error when using policy 0xd and 0cf. Fix tests 0521 and 0522 to
not expect IO errors. Also while at it, cleanup the test scripts
(including 0520 test) to quote variables and avoid long lines.

Signed-off-by: Damien Le Moal <damien.lemoal@opensource.wdc.com>