tests: test_lib: bump fio blocksize to 1280k
blocksize was recently increased to 1024k, since the min CDL timeout is
now 20ms.

Unfortunately, even with this larger blocksize, we occasionally still
see that some test cases do not trigger a CDL timeout.

Increase the blocksize to 1280k, which will hopefully make the command
execution take longer, such that the drive will always trigger a CDL
timeout.

If this is not enough, then we might need to set the max_sectors_kb
sysfs attribute to max_hw_sectors_kb, such that blocksize can be
increased even further.

Signed-off-by: Niklas Cassel <niklas.cassel@wdc.com>