tests: test_lib: make fio_get_clat_avg() more robust
While the logical thing would be for fio to always print the per prio
stats, regardless if fio exited successfully or not, that is not how
fio currently works.

Commit 79c144244b96 ("tests: test_lib: Improve cdl command percentage
and run time") started to print the average latency regardless if
compare_latencies was set or not, which caused fio_get_clat_avg() to
result in a syntax error, as the per prio stats are not in the fio
log file.

Make fio_get_clat_avg() return 0 if no per prio latencies are found,
and change analyze_log() to return an error on a zero latency.

Fixes: 79c144244b96 ("tests: test_lib: Improve cdl command percentage and run time")
Signed-off-by: Niklas Cassel <niklas.cassel@wdc.com>