tests: fail the test case if we detected a hard link reset in dmesg
Parse dmesg and see if a hard link reset occurred during the current
test case run, and if so, fail the test case.

A policy 0xF abort will return a NCQ error, however, even a NCQ error
should not cause the link to be hard reset.

Therefore, a hard reset is always a sign that something is seriously
wrong, so such a test case run should be marked as a failure.

This feature is similar to the CDL_REAL_RUNTIME check, which verifies
that the test case did not run longer than expected (which is usually
the case when there were I/O timeouts).

However, sometimes the link can be hard reset, and then return a CDL
timeout directly after the link has been reset, causing the total
runtime to still be less than the maximum allowed runtime. Those cases
are not caught by CDL_REAL_RUNTIME checks, so check dmesg as well.

Signed-off-by: Niklas Cassel <niklas.cassel@wdc.com>