tests: save dmesg snippet to ${tnum}_dmesg.log
It can be quite handy to have just the dmesg snippet that belongs to the
test case itself saved in cdl-test-logs, since running the whole test
suite can spam dmesg a lot, making it hard to find the correct output.

The dmesg log file can also be used to verify that the dmesg parsing is
working as intended.

Signed-off-by: Niklas Cassel <niklas.cassel@wdc.com>