tests: Improve run time to reduce false positives
Increase the runtime of fio to 60 seconds (regular run time) when errors
are expected to reduce the number of false positives results for cases
that expect IO errors. This is combined with using the
--continue_on_error=io option to continue executing fio even whe IO
errors are detected, unless quick tests were requested, in which case,
we request to stop on any error. fio ramp time is also modified and
requested only if IO errors are not expected.

Note: in practice, this does not increase significantly the tests run
time as fio continue_on_error option seems ineffective (likely an fio
bug).

Signed-off-by: Damien Le Moal <dlemoal@kernel.org>