tests: Fix test setup
While the page type to use for read and write commands is fixed for ATA
devices (T2A page for reads and T2B page for writes), SCSI devices can
have different associations between pages and command types.

Modify test_setup to correctly handle this with the new test lib
functions get_limits_page_type() and get_limits_page_file(). This last
helper generates the CDL page file to upload depending on the drive
page-to-command-type mapping and the command type specified (read or
write).

With this change, the interface of the execute_test() is modified,
replacing the T2A_file and T2B_file arguments with read_limits and
write_limits argumnets which are used to specify the name of the limits
to use for the command type. This name must match the CDL file names
under tests/scripts/cdl/, so that e.g. read_limits="XXX" leads to the
file tests/scripts/cdl/T2A-XXX.cdl or tests/scripts/cdl/T2B-XXX.cdl.

Signed-off-by: Damien Le Moal <dlemoal@kernel.org>