tests: Introduce test groups
Associate a test group with each test case. This is done by moving (and
for some cases re-numbering) test case scripts into sub-directories,
with each sub-directory corresponding to a test group. The name of each
directory indicates the group number and the group name.

Introduce the --group option to allow selecting all test cases of a
particular group. This option can be used multiple times to select
multiple groups. The logging directory short option "-g" is removed
reassigned as the short option of the --group option.

Signed-off-by: Damien Le Moal <dlemoal@kernel.org>