tests: Improve run time control options
Introduce the options --run-time and --stop-on-error to gain fine
control over fio execution, allowing to specify longer run times and to
ignore ETIME errors from aborted CDL commands.

The default value for the --run_time option is 60s, as before. The
--quick option is preserved and now implemented as an alias for
"--run-time 20 --stop-on-error".

This patch also fixes an issue with the use of fio --continue_on_error
option, which by default does not force ignoring the ETIME errors. This
is fixed by using fio option --ignore_error when --stop-on-error is
used. This fix implies that the error parsing done by fio_get_io_error()
must be improved to detect "Timer expired" error messages that are not
notified with an "err=" fatal error report message from fio.

Signed-off-by: Damien Le Moal <dlemoal@kernel.org>