tests: use t2cdlunits 0xa for an empty descriptor
The current SATL specification says that:
t2cdlunits is not changeable and has to be set to 0xa.

This is basically a direct opposite from SPC specification that says
that t2cdlunits 0x0 means that the descriptor is ignored.

A limit of 0 means unlimited, and a policy is only taken into account
once the limit has been exceeded.
So it should not matter if we use t2cdlunits: 0x0 or t2cdlunits: 0xa,
as long as the limit itself is 0.

Signed-off-by: Niklas Cassel <niklas.cassel@wdc.com>