Revert "tests: Use libaio for fio runs"
This reverts commit 71ec79c53d458bb58efef2900cdee2e7435892a7.

Systems that already had fio with --cmdprio_hint support installed
stops working with the error message:
"The installed version of fio does not seem to have CDL support
(--cmdprio_hint option not supported)"
with the introduction of this commit.

The reason is that for a fio binary compiled without the optional
libaio-devel dependency, trying to run fio with the libaio ioengine
will result in an error.

In fact, using io_uring does not require any additional user space
libraries (unlike libaio, which requires libaio-devel).
The only requirement for io_uring is a kernel version 5.1 or newer.

Support for CDL requires a kernel way more recent than 5.1 anyway.
The less dependencies the better, so revert back to using io_uring.

Signed-off-by: Niklas Cassel <niklas.cassel@wdc.com>