cdladm: Use ATA command to get minimum and maximum limits
SCSI does not define fields to discover a device minimum and maximum
time limits. In the case of an ATA device, these limits are available.
Introduce the function cdl_ata_get_limits() to obtain these limits and
use this function in cdl_scsi_init() so that the limits can be known
even for an ATA disk connected to a SAS HBA.

Signed-off-by: Damien Le Moal <dlemoal@kernel.org>