tools: update benchmark scripts
Rename the tools directory to "benchmark" and update the cdl_bench.sh
and other scripts to match the latest (upstream) fio and kernel
support for CDL.

A README.md file documenting how to use the cdl_bench.sh script and
how to process benchmark results is also added. A reference to this
documentation is added in cdl-tools main README.md file.

Signed-off-by: Damien Le Moal <dlemoal@kernel.org>