cdl: Define CDL statistics configuration and values
CDL statistics configuration and statistics values differ in ATA and
SCSI. With ATA, 2 selector bytes defined with each CDL descriptor define
two statistic values that can be retrieved through a log page. With
SCSI, the configuration allows enabling and disabling 4 different
statistics values with a log page.

Given that these definitions are not compatible and cannot be easily
unified under a single format, define a device cdl statistics
configuration and values as an union of ATA and SCSI statistics data
structures.

Signed-off-by: Damien Le Moal <dlemoal@kernel.org>