cdl: fix statistics handling
For ATA devices connected to a SAS HBA, CDL statistics operations are
executed using ata passthrough operations. To do so, we must ensure that
the device structure has an ata_cdl_log buffer to cache the ATA device
CDL log page, even when --force-ata is not used.

Do so by changing the declaration of ata_cdl_log to be a statically
allocated array within the device structure. Also introduce the helper
function cdl_ata_statistics_supported() to cleanup the code in cdladm.c
and to add statistics support checks where needed.

Signed-off-by: Damien Le Moal <dlemoal@kernel.org>