tests: 0705, 0706: Fix statistics and limits upload with SAS HBAs
When the test target drive is connected to a SAS HBA with a SAT
implementation on the HBA, the limits pages are not translated including
the ATA statistics selectors. This causes tests 0705 and 0706 to fail.
Correct this by running these tests only using --force-ata with the help
of the helper function dev_uses_hba_sat().

Signed-off-by: Damien Le Moal <dlemoal@kernel.org>