tests: Revalidate device after failure
If a test case fails generating many errors (e.g. timeouts), libata may
reduce the device link speed and disable NCQ for the device. This causes
the following test cases in the tests being executed to run with the
device not configured as it normally would be, e.g. with NCQ disabled.

To avoid this, use the function dev_revalidate() to remove and rescan a
device after a failed test. This restores the default (highest) link
speed for the device and NCQ feature.

Signed-off-by: Damien Le Moal <dlemoal@kernel.org>