tests: Add --size option to the test script
Sometimes it is necessary to run CDL tests not on the entire drive,
but on a lower part of the full sector range of the device. For
instance, this functionality can be very useful for testing zoned
CDL drives.

Add a new command line option to cdl-tests.sh script to specify the
number of sectors on the disk to be used for test i/o, starting from
the beginning of the device sector range. If this option is not
specified in the script command line, the entire device is exercised.

Signed-off-by: Dmitry Fomichev <dmitry.fomichev@wdc.com>
Signed-off-by: Damien Le Moal <dlemoal@kernel.org>