tests: Improve dev_revalidate()
Commit 858744785dfa ("tests: lib: Fix dev_revalidate()") fixed the
dev_revalidate() function to wait for the block device file to exist.
This is needed but not enough to ensure that a device revalidation
completed: while the block device file exist, the device capacity may be
0 and thus we need to wait for the device capacity to ne non zero to
ensure that the device revalidation completed.

Introduce the function dev_wait_for_revalidate() to check the device
capacity and make sure that this function fails after 30s waiting for
the device capacity to be valid. cdl-tests.sh is modified to stop
execution if dev_revalidate() fails, including failures of
dev_wait_for_revalidate().

Suggested-by: Dmitry Fomichev <dmitry.fomichev@wdc.com>
Fixes: 858744785dfa ("tests: lib: Fix dev_revalidate()")
Signed-off-by: Damien Le Moal <dlemoal@kernel.org>