tests: Relax statistics counter verification
Comparing the nuber of IO errors reported by fio to the statistics
counter values can lead to false-positives due to the fact that fio IOs
may be split and processed as 2 (or more) commands to the disk. This
sometimes leads to the statistics counter being larger than the number
of IO errors reported by fio. In addition, for tests using 2 different
statistics counters, we may not necessarilly always get errors for both
counters/IO types.

Avoid reporting errors for these cases by relaxing the checks to getting
at least one statistics counter to have errors for test cases using both
counters, and verify that the total number of errors reported with
statistics is at least as large as the number of errors reported by fio.
For these results, issue a warning instead of failing the test case.

Signed-off-by: Damien Le Moal <dlemoal@kernel.org>